+++ tests/ui/common/test_Edit.m
@@ -9,8 +9,11 @@ purge
 
 h = figure;
 
+fhDirect = @(src, evt)disp('direct callback');
+
 ui = mic.ui.common.Edit( ...
-    'cLabel', 'Saved pos', ...
+    'cLabel', 'Saved pos', ... 
+    'fhDirectCallback', fhDirect, ...
     'cType', 'd' ...
 );
 
@@ -21,4 +24,7 @@ addlistener(ui, 'eEnter', cbEnter);
 
 
 cbChange = @(src, evt) (fprintf('eChange callback ui.get() = %1.2f\n', ui.get()));
-addlistener(ui, 'eEnter', cbChange);
+addlistener(ui, 'eChange', cbChange);
+
+
+ui.set(2003)