+++ CHANGELOG.md
@@ -1,8 +1,20 @@
-﻿
-# 1.0.0-beta.25
+﻿# 1.0.0-beta.27
+
+### mic.ui.common.Edit/Button/Popup/PopupStruct/Checkbox
+
+Updated `fhDirectCallback` to have two args: `(src, evt)` to match standard callback pattern in MATLAB.  
+
+# 1.0.0-beta.26
+
+### mic.ui.common.Edit
+
+Updated mic.ui.common.Edit so it calls `fhDirectCallback()` any time it does `notify()`
+
+
 - Adding accessor for u8Index in `Scan.m`
 
-- Adding functionality to `ScanSetup` to allow parameter change callbacks.
+- Updating `mic.ui.axes.ScalableAxes.m` for bug fixes and customizable
+
 
 - Fixed a bug in `mic.scan` causing it to fail to stop when stop() is called during a state change or an acquisition. Also fixing a bug causing raster scanning to fail in `mic.ui.common.ScanSetup`
 
@@ -253,7 +265,7 @@ Added `save()` and `load()` methods and `saveToDisk()` and `loadFromDisk()` meth
 - Added `get()`, `getOptions()`, `getSelectedIndexes()`, `setOptions()` and `setSelectedIndexes()` methods
 - Fixed bug with `onMoveDown()` method when more than one item was selected while performing the action
 
-### mic.ui.common.Popup/Select
+### mic.ui.common.Popup*
 
 - Got rid of setters
 - Added  `get()`, `getOptions()`, `getSelectedIndex()`, `setOptions()` and `setSelectedIndex()` methods